requirements.txt
aniso8601==6.0.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.3
MarkupSafe==1.1.1
news-headline-scrapper==0.0.1
nltk==3.4
pox==0.2.5
ppft==1.6.4.9
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9
textblob==0.15.3
tornado==4.5.3
urllib3==1.24.1
Werkzeug==0.15.1
pandas == 0.24.1
numpy == 1.16.2
dtw == 1.3.3
matplotlib == 3.0.3
